backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.2",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.1",
    "web3>=7.3.0",
]
